python: Rework error handling of WrappingClient
If self.errormsgs is set to None for any code, that code is now handled
as a valid return.  If self.errormsgs is not None for any code, that
code is handled as an error, even if it would normally be a non-error
code (ie 200). This allows for testing for existence of a resource
before overwriting it.